:recycle: Moved Select.test tests to ComboBox.test/SingleSelect.test
