+++ src/molecules/Select/Select.test.tsx
@@ -11,7 +11,6 @@ import {
   render,
   screen,
   userEvent,
-  within,
 } from 'src/tests/test-utils';
 
 import { expect } from 'vitest';
@@ -29,7 +28,7 @@ export function fakeGroups(count = 5, isParented?: boolean) {
     );
 }
 
-function fakeItemsWithChildren(count = 5) {
+export function fakeItemsWithChildren(count = 5) {
   return structuredClone(
     new Array(count).fill(0).map(() => ({
       ...fakeSelectItem(),
@@ -37,484 +36,6 @@ function fakeItemsWithChildren(count = 5) {
     }))
   );
 }
-
-test('Basic single select', async () => {
-  const items = fakeSelectItems();
-  const label = faker.animal.bear();
-  const meta = faker.animal.insect();
-  const handleOnSelect = vi.fn();
-  const { rerender } = render(
-    <Select
-      label={label}
-      meta={meta}
-      onSelect={handleOnSelect}
-      value={undefined}
-      items={items}
-    />
-  );
-  const user = userEvent.setup();
-
-  expect(screen.getByText(label)).toBeInTheDocument();
-  expect(screen.getByText(meta)).toBeInTheDocument();
-
-  await user.click(screen.getByRole('combobox'));
-
-  for (const item of items) {
-    expect(screen.getByText(item.label)).toBeInTheDocument();
-  }
-
-  const randomItem = faker.helpers.arrayElement(items);
-
-  await user.click(screen.getByText(randomItem.label));
-
-  expect(handleOnSelect).toHaveBeenCalledTimes(1);
-  expect(handleOnSelect).toHaveBeenCalledWith(randomItem);
-
-  rerender(
-    <Select
-      label={label}
-      meta={meta}
-      onSelect={handleOnSelect}
-      value={randomItem}
-      items={items}
-    />
-  );
-});
-
-test('Basic single select with helperText', () => {
-  const items = fakeSelectItems();
-  const helperText = faker.airline.airplane().name;
-  const handleOnSelect = vi.fn();
-  render(
-    <Select
-      helperText={helperText}
-      onSelect={handleOnSelect}
-      value={undefined}
-      items={items}
-    />
-  );
-
-  expect(screen.getByText(helperText)).toBeInTheDocument();
-});
-
-test('Basic single select with helperText and variant', () => {
-  const items = fakeSelectItems();
-  const helperText = faker.airline.airplane().name;
-  const handleOnSelect = vi.fn();
-  render(
-    <Select
-      helperText={helperText}
-      onSelect={handleOnSelect}
-      value={undefined}
-      items={items}
-      variant="error"
-    />
-  );
-
-  const helperTextElement = screen.getByText(helperText);
-  const helperIcon = within(
-    helperTextElement.parentElement!.parentElement!
-  ).getByTestId('eds-icon-path');
-
-  expect(helperTextElement).toBeInTheDocument();
-  expect(helperIcon).toBeInTheDocument();
-});
-
-test('Basic single select with only meta label', () => {
-  const items = fakeSelectItems();
-  const meta = faker.animal.insect();
-  const handleOnSelect = vi.fn();
-  render(
-    <Select
-      meta={meta}
-      onSelect={handleOnSelect}
-      value={undefined}
-      items={items}
-    />
-  );
-
-  expect(screen.getByText(meta)).toBeInTheDocument();
-});
-
-test('Basic multi select', async () => {
-  const items = fakeSelectItems();
-  const label = faker.animal.bear();
-  const handleOnSelect = vi.fn();
-  render(
-    <Select label={label} onSelect={handleOnSelect} values={[]} items={items} />
-  );
-  const user = userEvent.setup();
-
-  await user.click(screen.getByRole('combobox'));
-
-  for (const item of items) {
-    expect(screen.getByText(item.label)).toBeInTheDocument();
-  }
-
-  const randomItems = faker.helpers.arrayElements(items);
-
-  for (const item of randomItems) {
-    await user.click(screen.getByText(item.label));
-    expect(handleOnSelect).toHaveBeenCalledWith([item], item);
-  }
-
-  expect(handleOnSelect).toHaveBeenCalledTimes(randomItems.length);
-});
-
-test('Parent multi select with selectableParent = false', async () => {
-  const items = fakeItemsWithChildren();
-  const label = faker.animal.bear();
-  const handleOnSelect = vi.fn();
-  const { rerender } = render(
-    <Select
-      label={label}
-      onSelect={handleOnSelect}
-      values={[]}
-      items={items}
-      selectableParent={false}
-    />
-  );
-  const user = userEvent.setup();
-
-  await user.click(screen.getByRole('combobox'));
-
-  for (const item of items) {
-    expect(screen.getByText(item.label)).toBeInTheDocument();
-  }
-
-  const randomItem = faker.helpers.arrayElement(items);
-
-  await user.click(screen.getByText(randomItem.label));
-
-  expect(handleOnSelect).toHaveBeenCalledTimes(1);
-  handleOnSelect.mockClear();
-
-  for (const child of randomItem.children) {
-    const menuItem = screen.getByRole('menuitem', { name: child.label });
-    await user.click(menuItem);
-    expect(handleOnSelect).toHaveBeenCalledWith([child], child);
-  }
-
-  expect(handleOnSelect).toHaveBeenCalledTimes(randomItem.children.length);
-
-  rerender(
-    <Select
-      label={label}
-      onSelect={handleOnSelect}
-      values={[randomItem.children[0]]}
-      items={items}
-      selectableParent={false}
-    />
-  );
-
-  await user.click(
-    screen.getByRole('menuitem', { name: randomItem.children[0].label })
-  );
-  expect(handleOnSelect).toHaveBeenCalledWith([], randomItem.children[0]);
-});
-
-test('Parent multi select with selectableParent = true', async () => {
-  const items = fakeItemsWithChildren();
-  const itemWithoutChildren = fakeSelectItem();
-  const label = faker.animal.bear();
-  const handleOnSelect = vi.fn();
-  const { rerender } = render(
-    <Select
-      label={label}
-      onSelect={handleOnSelect}
-      values={[]}
-      items={[...items, itemWithoutChildren]}
-      selectableParent
-    />
-  );
-  const user = userEvent.setup();
-
-  await user.click(screen.getByRole('combobox'));
-
-  for (const item of items) {
-    expect(screen.getByText(item.label)).toBeInTheDocument();
-  }
-
-  const randomItem = faker.helpers.arrayElement(items);
-
-  const icon = screen.getByRole('menuitem', { name: randomItem.label });
-  await user.click(icon);
-
-  expect(handleOnSelect).toHaveBeenCalledWith([randomItem], randomItem);
-
-  rerender(
-    <Select
-      label={label}
-      onSelect={handleOnSelect}
-      values={[randomItem]}
-      items={[...items, itemWithoutChildren]}
-    />
-  );
-
-  await user.click(icon);
-  expect(handleOnSelect).toHaveBeenCalledWith([], randomItem);
-
-  rerender(
-    <Select
-      label={label}
-      onSelect={handleOnSelect}
-      values={[itemWithoutChildren]}
-      items={[...items, itemWithoutChildren]}
-    />
-  );
-
-  await user.click(
-    screen.getByRole('menuitem', { name: itemWithoutChildren.label })
-  );
-  expect(handleOnSelect).toHaveBeenCalledWith([], itemWithoutChildren);
-});
-
-test('Parent multi select - nested selection works as expected with keyboard', async () => {
-  const label = faker.animal.bear();
-  const items = fakeItemsWithChildren();
-  const handler = vi.fn();
-
-  const { rerender } = render(
-    <Select label={label} onSelect={handler} items={items} values={[]} />
-  );
-  const user = userEvent.setup({ delay: 100 });
-
-  await user.click(screen.getByRole('combobox'));
-
-  // Enter menu
-  await user.keyboard('{ArrowDown}');
-  await user.keyboard('{ArrowRight}');
-
-  await user.keyboard('{ArrowDown}');
-
-  await user.keyboard('{Enter}');
-  let newValues = [items[0].children[0]];
-  expect(handler).toBeCalledWith(newValues, items[0].children[0]);
-
-  rerender(
-    <Select label={label} onSelect={handler} items={items} values={newValues} />
-  );
-
-  const icon1 = screen.getByRole('menuitem', {
-    name: items[0].children[0].label,
-  });
-
-  const icon2 = screen.getByRole('menuitem', {
-    name: items[0].children[1].label,
-  });
-
-  await user.click(icon2);
-
-  newValues = [items[0].children[1], items[0].children[0]];
-  expect(handler).toBeCalledWith(newValues, items[0].children[1]);
-
-  rerender(
-    <Select label={label} onSelect={handler} items={items} values={newValues} />
-  );
-
-  await user.click(icon2);
-
-  newValues = [items[0].children[0]];
-  expect(handler).toBeCalledWith(newValues, items[0].children[1]);
-
-  rerender(
-    <Select label={label} onSelect={handler} items={items} values={newValues} />
-  );
-
-  await user.click(icon1);
-
-  expect(handler).toBeCalledWith([], items[0].children[0]);
-});
-
-test('Parent multi select - nested selection with preselected parent works as expected', async () => {
-  const label = faker.animal.bear();
-  const items = fakeItemsWithChildren();
-  const handler = vi.fn();
-
-  render(
-    <Select
-      label={label}
-      onSelect={handler}
-      items={items}
-      values={[items[0]]}
-    />
-  );
-  const user = userEvent.setup({ delay: 100 });
-
-  await user.click(screen.getByRole('combobox'));
-
-  // Go past and back  -- for some reason if you remove the next two lines the test sometimes fails 😅
-  await user.keyboard('{ArrowDown}');
-  await user.keyboard('{ArrowUp}');
-
-  // Enter menu
-  await user.keyboard('{ArrowDown}');
-  await user.keyboard('{ArrowRight}');
-
-  await user.keyboard('{ArrowDown}');
-  await user.keyboard('{Enter}');
-  expect(handler).toBeCalledWith([items[0].children[0]], items[0].children[0]);
-});
-
-test('Parent multi select - nested child label shows as expected', async () => {
-  const label = faker.animal.bear();
-  const items = [
-    {
-      label: 'label 1',
-      value: 'value 1',
-      children: [
-        {
-          label: 'label 2',
-          value: 'value 2',
-          children: [
-            {
-              label: 'label 3',
-              value: 'value 3',
-            },
-          ],
-        },
-      ],
-    },
-  ];
-
-  const handler = vi.fn();
-
-  render(
-    <Select
-      label={label}
-      onSelect={handler}
-      items={items}
-      values={[{ ...items[0].children[0].children[0], children: [] }]}
-    />
-  );
-
-  const user = userEvent.setup();
-  await user.click(screen.getByRole('combobox'));
-
-  expect(screen.getByText('label 3')).toBeInTheDocument();
-});
-
-test('Basic group single select', async () => {
-  const label = faker.animal.bear();
-  const handler = vi.fn();
-  const groups = fakeGroups();
-  render(
-    <Select
-      label={label}
-      onSelect={handler}
-      groups={groups}
-      value={undefined}
-    />
-  );
-
-  const user = userEvent.setup();
-
-  await user.click(screen.getByRole('combobox'));
-
-  for (const group of groups) {
-    expect(screen.getByText(group.title)).toBeInTheDocument();
-    for (const item of group.items) {
-      expect(screen.getByText(item.label)).toBeInTheDocument();
-    }
-  }
-
-  const randomItem = faker.helpers.arrayElement(
-    faker.helpers.arrayElement(groups).items
-  );
-
-  await user.click(screen.getByText(randomItem.label));
-
-  expect(handler).toBeCalledTimes(1);
-  expect(handler).toBeCalledWith(randomItem);
-});
-
-test('Basic group multi select with preselected item', async () => {
-  const label = faker.animal.bear();
-  const handler = vi.fn();
-  const groups = fakeGroups();
-
-  const randomItem = faker.helpers.arrayElement(
-    faker.helpers.arrayElement(groups).items
-  );
-  render(
-    <Select
-      label={label}
-      onSelect={handler}
-      groups={groups}
-      values={[randomItem]}
-    />
-  );
-
-  const user = userEvent.setup();
-
-  await user.click(screen.getByRole('combobox'));
-
-  await user.click(screen.getByRole('menuitem', { name: randomItem.label }));
-
-  expect(handler).toBeCalledTimes(1);
-  expect(handler).toBeCalledWith([], randomItem);
-});
-
-test('Parented group multi select', async () => {
-  const label = faker.animal.bear();
-  const handler = vi.fn();
-  const groups = fakeGroups(5, true);
-
-  render(
-    <Select label={label} onSelect={handler} groups={groups} values={[]} />
-  );
-
-  const user = userEvent.setup();
-
-  await user.click(screen.getByRole('combobox'));
-
-  const firstItem = groups[0].items[0];
-
-  await user.click(screen.getByRole('menuitem', { name: firstItem.label }));
-
-  expect(handler).toBeCalledTimes(1);
-  expect(handler).toBeCalledWith([firstItem], firstItem);
-});
-
-test('Sorts items as expected', () => {
-  const label = faker.animal.bear();
-  const items = fakeSelectItems();
-  const values = [items[3], items[1], items[0]];
-  const sortedValues = [items[0], items[1], items[3]];
-  const handler = vi.fn();
-
-  const { rerender } = render(
-    <Select label={label} onSelect={handler} items={items} values={values} />
-  );
-
-  for (let i = 1; i < sortedValues.length; i++) {
-    const first = screen.getByText(sortedValues[i - 1].label);
-    const second = screen.getByText(sortedValues[i].label);
-
-    expect(first.compareDocumentPosition(second)).toBe(
-      Node.DOCUMENT_POSITION_FOLLOWING
-    );
-  }
-
-  rerender(
-    <Select
-      label={label}
-      onSelect={handler}
-      items={items}
-      values={values}
-      sortValues={false}
-    />
-  );
-
-  for (let i = 1; i < values.length; i++) {
-    const first = screen.getByText(values[i - 1].label);
-    const second = screen.getByText(values[i].label);
-
-    expect(first.compareDocumentPosition(second)).toBe(
-      Node.DOCUMENT_POSITION_FOLLOWING
-    );
-  }
-});
 
 test('Can open/close by clicking icon', async () => {
   const label = faker.animal.bear();
@@ -631,32 +152,6 @@ test("Clicking 'x' on chip works as expected", async () => {
   );
 
   await user.click(chip);
-
-  expect(handler).toHaveBeenCalledWith([], items[0]);
-});
-
-test('Removing with backspace', async () => {
-  const label = faker.animal.bear();
-  const items = fakeSelectItems();
-  const handler = vi.fn();
-
-  render(
-    <Select
-      label={label}
-      onSelect={handler}
-      items={items}
-      values={[items[0]]}
-    />
-  );
-  const user = userEvent.setup();
-
-  const searchField = screen.getByRole('combobox');
-
-  await user.click(searchField);
-
-  await user.keyboard('{Backspace}');
-
-  await user.keyboard('{Backspace}');
 
   expect(handler).toHaveBeenCalledWith([], items[0]);
 });
